refactor(registration): split RegistrationForm into Public/Invite variants
Replaces the single ~1400-line RegistrationForm + cascade of mode booleans
(`inviteMode`, `registrationOpen`, `registrationAllowed`) with two explicit
top-level variants per Vercel composition guidelines: PublicRegistrationForm
and InviteRegistrationForm. Both compose a shared set of section components
(DivisionPickerSection, AffiliateSection, FeeSummarySection, etc.) and pull
state from a single useRegistrationForm hook. Each variant is intrinsically
clear about what it renders — no abstract "mode" threaded through children,
no `!registrationOpen || inviteMode` gymnastics.

Invite variant adds a celebratory InviteDivisionHero that swaps in for the
multi-select picker: blurred competition profile image background, conic
gradient ring, gradient division reveal. Invitations are a high-emotion
moment for athletes; a disabled dropdown undersells it. The hero pins the
selection to the invited division for the rest of the form. Falls back to
the public variant when the invited division becomes ineligible.

Route dispatches on URL `?divisionId=<x>` (the only producer is the claim
CTA). Claim landing now forwards `invite=<token>` alongside divisionId so
the registration form can present authorization to the server.
