+++ apps/wodsmith-start/src/components/registration/registration-form.tsx
@@ -1,46 +1,7 @@
 import { useNavigate } from "@tanstack/react-router"
-import { useServerFn } from "@tanstack/react-start"
-import {
-  AlertTriangle,
-  Check,
-  CheckCircle2,
-  ChevronsUpDown,
-  Loader2,
-  Search,
-  Tag,
-  User,
-  Users,
-  X,
-} from "lucide-react"
-import { useEffect, useRef, useState } from "react"
-import { toast } from "sonner"
-import { trackEvent } from "@/lib/posthog"
-import { clearCouponSession, getCouponSession } from "@/utils/coupon-cookie"
-import { WaiverViewer } from "@/components/compete/waiver-viewer"
-import { Badge } from "@/components/ui/badge"
+import { Loader2 } from "lucide-react"
+import { InviteDivisionHero } from "@/components/registration/invite-division-hero"
 import { Button } from "@/components/ui/button"
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card"
-import { Checkbox } from "@/components/ui/checkbox"
-import { Input } from "@/components/ui/input"
-import { Label } from "@/components/ui/label"
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from "@/components/ui/popover"
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select"
 import type {
   Competition,
   ScalingGroup,
@@ -49,209 +10,37 @@ import type {
   Waiver,
 } from "@/db/schema"
 import type { PublicCompetitionDivision } from "@/server-fns/competition-divisions-fns"
-import type { CompetitionCapacityResult } from "@/utils/competition-capacity"
-import { initiateRegistrationPaymentFn } from "@/server-fns/registration-fns"
 import type { RegistrationQuestion } from "@/server-fns/registration-questions-fns"
-import { signWaiverFn } from "@/server-fns/waiver-fns"
-import { cn } from "@/utils/cn"
-import { getLocalDateKey, isSameDateString } from "@/utils/date-utils"
-import { AffiliateCombobox } from "./affiliate-combobox"
-import { FeeBreakdown } from "./fee-breakdown"
+import type { CompetitionCapacityResult } from "@/utils/competition-capacity"
+import {
+  AffiliateSection,
+  CapacityBanners,
+  ClosedRegistrationBanner,
+  CompetitionDetailsCard,
+  DivisionPickerSection,
+  FeeSummarySection,
+  PageHeader,
+  RegisteredAlert,
+  RegistrationQuestionsSection,
+  RemovedAlert,
+  TeamDetailsSection,
+  WaiversSection,
+} from "./registration-sections"
+import {
+  type UseRegistrationFormReturn,
+  useRegistrationForm,
+} from "./use-registration-form"
 
-interface Teammate {
-  email: string
-  firstName: string
-  lastName: string
-  affiliateName: string
-}
-
-interface TeamEntry {
-  divisionId: string
-  teamName: string
-  teammates: Teammate[]
-}
-
-function DivisionMultiSelect({
-  scalingLevels,
-  publicDivisions,
-  selectedIds,
-  registeredDivisionIds,
-  removedDivisionIds,
-  onToggle,
-  disabled,
-}: {
-  scalingLevels: ScalingLevel[]
-  publicDivisions: PublicCompetitionDivision[]
-  selectedIds: string[]
-  registeredDivisionIds: Set<string>
-  removedDivisionIds: Set<string>
-  onToggle: (id: string, checked: boolean) => void
-  disabled: boolean
-}) {
-  const [open, setOpen] = useState(false)
-  const [search, setSearch] = useState("")
-  const inputRef = useRef<HTMLInputElement>(null)
-
-  const filtered = scalingLevels.filter((l) =>
-    l.label.toLowerCase().includes(search.toLowerCase()),
-  )
-
-  // Focus search input when popover opens
-  useEffect(() => {
-    if (open) {
-      setTimeout(() => inputRef.current?.focus(), 0)
-    } else {
-      setSearch("")
-    }
-  }, [open])
-
-  const getPublicDiv = (id: string) => publicDivisions.find((d) => d.id === id)
-
-  return (
-    <Popover open={open} onOpenChange={setOpen}>
-      <PopoverTrigger asChild>
-        {/* biome-ignore lint/a11y/useSemanticElements: Custom combobox pattern */}
-        <Button
-          variant="outline"
-          role="combobox"
-          aria-expanded={open}
-          className="w-full justify-between font-normal"
-          disabled={disabled}
-        >
-          <span className="truncate text-muted-foreground">
-            {selectedIds.length === 0
-              ? "Search divisions..."
-              : `${selectedIds.length} division${selectedIds.length > 1 ? "s" : ""} selected`}
-          </span>
-          <ChevronsUpDown className="ml-2 h-4 w-4 shrink-0 opacity-50" />
-        </Button>
-      </PopoverTrigger>
-      <PopoverContent
-        className="w-[var(--radix-popover-trigger-width)] p-0"
-        align="start"
-      >
-        <div className="flex items-center border-b px-3 py-2">
-          <Search className="mr-2 h-4 w-4 shrink-0 opacity-50" />
-          <Input
-            ref={inputRef}
-            placeholder="Search divisions..."
-            value={search}
-            onChange={(e) => setSearch(e.target.value)}
-            className="h-8 border-0 p-0 placeholder:text-muted-foreground focus-visible:ring-0 focus-visible:ring-offset-0"
-          />
-        </div>
-        <div className="max-h-[300px] overflow-y-auto p-1">
-          {filtered.length === 0 ? (
-            <p className="py-6 text-center text-sm text-muted-foreground">
-              No divisions found.
-            </p>
-          ) : (
-            filtered.map((level) => {
-              const info = getPublicDiv(level.id)
-              const isFull = info?.isFull ?? false
-              const spotsAvailable = info?.spotsAvailable
-              const maxSpots = info?.maxSpots
-              const isRegistered = registeredDivisionIds.has(level.id)
-              const isRemoved = removedDivisionIds.has(level.id)
-              const isSelected = selectedIds.includes(level.id)
-              const isItemDisabled =
-                disabled || isFull || isRegistered || isRemoved
-
-              return (
-                <button
-                  key={level.id}
-                  type="button"
-                  onClick={() => {
-                    if (!isItemDisabled) {
-                      onToggle(level.id, !isSelected)
-                    }
-                  }}
-                  disabled={isItemDisabled}
-                  className={cn(
-                    "flex w-full items-center gap-2 rounded-sm px-2 py-2 text-sm outline-none hover:bg-accent hover:text-accent-foreground",
-                    isSelected && "bg-accent/50",
-                    isItemDisabled &&
-                      "opacity-50 cursor-not-allowed hover:bg-transparent",
-                  )}
-                >
-                  <Check
-                    className={cn(
-                      "h-4 w-4 shrink-0",
-                      isSelected ? "opacity-100" : "opacity-0",
-                    )}
-                  />
-                  <span
-                    className={cn(
-                      "flex-1 text-left font-medium",
-                      isFull &&
-                        !isRegistered &&
-                        "line-through text-muted-foreground",
-                    )}
-                  >
-                    {level.label}
-                  </span>
-                  <div className="flex items-center gap-1.5 shrink-0">
-                    {(level.teamSize ?? 1) > 1 ? (
-                      <Badge variant="secondary" className="text-xs">
-                        <Users className="w-3 h-3 mr-1" />
-                        {level.teamSize}
-                      </Badge>
-                    ) : (
-                      <Badge variant="outline" className="text-xs">
-                        <User className="w-3 h-3 mr-1" />
-                        Indy
-                      </Badge>
-                    )}
-                    {isRemoved ? (
-                      <Badge
-                        variant="outline"
-                        className="text-xs text-destructive border-destructive/30"
-                      >
-                        Removed
-                      </Badge>
-                    ) : isRegistered ? (
-                      <Badge
-                        variant="outline"
-                        className="text-xs text-green-600 border-green-500/30"
-                      >
-                        <CheckCircle2 className="w-3 h-3 mr-1" />
-                        Registered
-                      </Badge>
-                    ) : isFull ? (
-                      <Badge variant="destructive" className="text-xs">
-                        SOLD OUT
-                      </Badge>
-                    ) : maxSpots !== null &&
-                      spotsAvailable !== null &&
-                      spotsAvailable !== undefined &&
-                      spotsAvailable <= 5 ? (
-                      <Badge
-                        variant="secondary"
-                        className="text-xs text-amber-600 dark:text-amber-400"
-                      >
-                        {spotsAvailable} left
-                      </Badge>
-                    ) : null}
-                  </div>
-                </button>
-              )
-            })
-          )}
-        </div>
-      </PopoverContent>
-    </Popover>
-  )
-}
-
-type Props = {
+// Loader-data shape shared by both variants. Kept colocated so the route can
+// pass the same payload regardless of which variant it picks.
+export interface RegistrationFormProps {
   competition: Competition & { organizingTeam: Team | null }
   scalingGroup: ScalingGroup & { scalingLevels: ScalingLevel[] }
   publicDivisions: PublicCompetitionDivision[]
   competitionCapacity?: CompetitionCapacityResult | null
   userId: string
-  registrationOpen: boolean
-  registrationOpensAt: string | null // YYYY-MM-DD format
-  registrationClosesAt: string | null // YYYY-MM-DD format
+  registrationOpensAt: string | null
+  registrationClosesAt: string | null
   paymentCanceled?: boolean
   defaultAffiliateName?: string
   waivers: Waiver[]
@@ -265,1073 +54,118 @@ type Props = {
   signedWaiverIds?: string[]
 }
 
-export function RegistrationForm({
-  competition,
-  scalingGroup,
-  publicDivisions,
-  competitionCapacity,
-  userId: _userId,
-  registrationOpen,
-  registrationOpensAt,
-  registrationClosesAt,
-  paymentCanceled,
-  defaultAffiliateName,
-  waivers,
-  questions,
-  userFirstName,
-  userLastName,
-  userEmail,
-  registeredDivisionIds = [],
-  removedDivisionIds = [],
-  previousAnswers = [],
-  signedWaiverIds = [],
-}: Props) {
+interface PublicProps extends RegistrationFormProps {
+  registrationOpen: boolean
+}
+
+interface InviteProps extends RegistrationFormProps {
+  initialDivisionId: string
+  inviteToken?: string
+}
+
+// ─── Public variant ─────────────────────────────────────────────────────────
+
+export function PublicRegistrationForm(props: PublicProps) {
+  const r = useRegistrationForm(props)
   const navigate = useNavigate()
-  const [isSubmitting, setIsSubmitting] = useState(false)
-
-  // Active coupon from sessionStorage
-  const [activeCoupon] = useState(() => {
-    const coupon = getCouponSession()
-    return coupon?.competitionSlug === competition.slug ? coupon : null
-  })
-
-  // Multi-select state
-  const [selectedDivisionIds, setSelectedDivisionIds] = useState<string[]>([])
-  const [affiliateName, setAffiliateName] = useState(defaultAffiliateName ?? "")
-
-  // Team details per division (for team divisions)
-  const [teamEntries, setTeamEntries] = useState<Map<string, TeamEntry>>(
-    new Map(),
-  )
-
-  // Fee data per division (for combined total)
-  const [divisionFees, setDivisionFees] = useState<Map<string, number>>(
-    new Map(),
-  )
-
-  // Prune fee entries for deselected divisions
-  useEffect(() => {
-    setDivisionFees((prev) => {
-      const selectedSet = new Set(selectedDivisionIds)
-      let changed = false
-      for (const key of prev.keys()) {
-        if (!selectedSet.has(key)) {
-          changed = true
-        }
-      }
-      if (!changed) return prev
-      const next = new Map<string, number>()
-      for (const [k, v] of prev) {
-        if (selectedSet.has(k)) next.set(k, v)
-      }
-      return next
-    })
-  }, [selectedDivisionIds])
-
-  const handleFeesLoaded = (
-    divisionId: string,
-    fees: { isFree: boolean; totalChargeCents?: number } | null,
-  ) => {
-    setDivisionFees((prev) => {
-      const next = new Map(prev)
-      if (fees && !fees.isFree && fees.totalChargeCents) {
-        next.set(divisionId, fees.totalChargeCents)
-      } else {
-        next.delete(divisionId)
-      }
-      return next
-    })
-  }
-
-  // Registration question answers - prefill from previous registration if available
-  const [answers, setAnswers] = useState<
-    Array<{ questionId: string; answer: string }>
-  >(
-    questions.map((q) => {
-      const prev = previousAnswers.find((a) => a.questionId === q.id)
-      return { questionId: q.id, answer: prev?.answer ?? "" }
-    }),
-  )
-
-  // Track which waivers have been agreed to - prefill from previous signatures
-  const [agreedWaivers, setAgreedWaivers] = useState<Set<string>>(() => {
-    const signedSet = new Set(signedWaiverIds)
-    const preAgreed = new Set<string>()
-    for (const w of waivers) {
-      if (signedSet.has(w.id)) {
-        preAgreed.add(w.id)
-      }
-    }
-    return preAgreed
-  })
-
-  // Use useServerFn for TanStack Start pattern
-  const signWaiver = useServerFn(signWaiverFn)
-
-  // Check if all required waivers are agreed to
-  const requiredWaivers = waivers.filter((w) => w.required)
-  const allRequiredWaiversAgreed = requiredWaivers.every((w) =>
-    agreedWaivers.has(w.id),
-  )
-
-  const registeredDivisionIdSet = new Set(registeredDivisionIds)
-  const removedDivisionIdSet = new Set(removedDivisionIds)
-
-  // Track registration started on mount
-  useEffect(() => {
-    trackEvent("competition_registration_started", {
-      competition_id: competition.id,
-      competition_name: competition.name,
-      competition_slug: competition.slug,
-    })
-  }, [competition.id, competition.name, competition.slug])
-
-  // Show toast if returning from canceled payment
-  useEffect(() => {
-    if (paymentCanceled) {
-      toast.error("Payment was canceled. Please try again when you're ready.")
-    }
-  }, [paymentCanceled])
-
-  // Helper to get division info
-  const getDivision = (divisionId: string) =>
-    scalingGroup.scalingLevels.find((l) => l.id === divisionId)
-
-  // Handle division checkbox toggle
-  const handleDivisionToggle = (divisionId: string, checked: boolean) => {
-    if (checked) {
-      setSelectedDivisionIds((prev) => [...prev, divisionId])
-      const division = getDivision(divisionId)
-      if (division && division.teamSize > 1) {
-        // Initialize team entry for team division
-        const teammatesNeeded = division.teamSize - 1
-        setTeamEntries((prev) => {
-          const next = new Map(prev)
-          next.set(divisionId, {
-            divisionId,
-            teamName: "",
-            teammates: Array.from({ length: teammatesNeeded }, () => ({
-              email: "",
-              firstName: "",
-              lastName: "",
-              affiliateName: "",
-            })),
-          })
-          return next
-        })
-      }
-    } else {
-      setSelectedDivisionIds((prev) => prev.filter((id) => id !== divisionId))
-      setTeamEntries((prev) => {
-        const next = new Map(prev)
-        next.delete(divisionId)
-        return next
-      })
-    }
-  }
-
-  // Update team entry field
-  const updateTeamEntry = (
-    divisionId: string,
-    field: "teamName",
-    value: string,
-  ) => {
-    setTeamEntries((prev) => {
-      const next = new Map(prev)
-      const entry = next.get(divisionId)
-      if (entry) {
-        next.set(divisionId, { ...entry, [field]: value })
-      }
-      return next
-    })
-  }
-
-  const updateTeammate = (
-    divisionId: string,
-    index: number,
-    field: keyof Teammate,
-    value: string,
-  ) => {
-    setTeamEntries((prev) => {
-      const next = new Map(prev)
-      const entry = next.get(divisionId)
-      if (entry) {
-        const teammates = [...entry.teammates]
-        teammates[index] = { ...teammates[index], [field]: value }
-        next.set(divisionId, { ...entry, teammates })
-      }
-      return next
-    })
-  }
-
-  const handleWaiverCheckChange = (waiverId: string, checked: boolean) => {
-    setAgreedWaivers((prev) => {
-      const newSet = new Set(prev)
-      if (checked) {
-        newSet.add(waiverId)
-      } else {
-        newSet.delete(waiverId)
-      }
-      return newSet
-    })
-  }
-
-  const updateAnswer = (questionId: string, value: string) => {
-    setAnswers((prev) =>
-      prev.map((a) =>
-        a.questionId === questionId ? { ...a, answer: value } : a,
-      ),
-    )
-  }
-
-  // Build items for multi-division submission
-  const buildRegistrationItems = () => {
-    return selectedDivisionIds.map((divisionId) => {
-      const division = getDivision(divisionId)
-      const isTeam = (division?.teamSize ?? 1) > 1
-      const teamEntry = teamEntries.get(divisionId)
-
-      return {
-        divisionId,
-        teamName: isTeam ? teamEntry?.teamName : undefined,
-        teammates: isTeam ? teamEntry?.teammates : undefined,
-      }
-    })
-  }
-
-  const onSubmit = async (e: React.FormEvent) => {
-    e.preventDefault()
-
-    if (selectedDivisionIds.length === 0) {
-      toast.error("Please select at least one division")
-      return
-    }
-
-    if (!affiliateName.trim()) {
-      toast.error("Please select your affiliate or Independent")
-      return
-    }
-
-    // Validate team fields for each team division
-    for (const divisionId of selectedDivisionIds) {
-      const division = getDivision(divisionId)
-      if (!division || division.teamSize <= 1) continue
-
-      const teamEntry = teamEntries.get(divisionId)
-      if (!teamEntry?.teamName?.trim()) {
-        toast.error(`Team name is required for ${division.label}`)
-        return
-      }
-
-      const teammatesNeeded = division.teamSize - 1
-      if (
-        !teamEntry.teammates ||
-        teamEntry.teammates.length !== teammatesNeeded
-      ) {
-        toast.error(
-          `Please add ${teammatesNeeded} teammate(s) for ${division.label}`,
-        )
-        return
-      }
-
-      for (const teammate of teamEntry.teammates) {
-        if (!teammate.email?.trim()) {
-          toast.error(`All teammate emails are required for ${division.label}`)
-          return
-        }
-      }
-    }
-
-    // Validate required registration questions
-    if (questions.length > 0) {
-      for (const question of questions) {
-        if (question.required) {
-          const answer = answers.find((a) => a.questionId === question.id)
-          if (!answer?.answer?.trim()) {
-            toast.error(
-              `Please answer the required question: ${question.label}`,
-            )
-            return
-          }
-        }
-      }
-    }
-
-    // Check waivers
-    if (!allRequiredWaiversAgreed) {
-      toast.error("Please agree to all required waivers before registering")
-      return
-    }
-
-    setIsSubmitting(true)
-
-    try {
-      // Sign all agreed waivers first
-      for (const waiverId of agreedWaivers) {
-        const result = await signWaiver({
-          data: {
-            waiverId,
-            registrationId: undefined,
-            ipAddress: undefined,
-          },
-        })
-
-        if (!result.success) {
-          toast.error("Failed to sign waiver")
-          setIsSubmitting(false)
-          return
-        }
-      }
-
-      // Build registration items
-      const items = buildRegistrationItems()
-
-      // Submit registration
-      const result = await initiateRegistrationPaymentFn({
-        data: {
-          competitionId: competition.id,
-          items,
-          affiliateName: affiliateName || undefined,
-          answers,
-          couponCode: activeCoupon?.code,
-        },
-      })
-
-      // FREE registration - redirect to registered page
-      if (result.isFree) {
-        if (activeCoupon) clearCouponSession()
-        trackEvent("competition_registration_completed", {
-          competition_id: competition.id,
-          competition_name: competition.name,
-          competition_slug: competition.slug,
-          division_count: items.length,
-        })
-        toast.success("Successfully registered!")
-        navigate({
-          to: `/compete/${competition.slug}/registered`,
-          search: { registration_id: result.registrationId ?? undefined },
-        })
-        return
-      }
-
-      // PAID registration - redirect to Stripe Checkout
-      if (result.checkoutUrl) {
-        if (activeCoupon) clearCouponSession()
-        trackEvent("competition_registration_payment_started", {
-          competition_id: competition.id,
-          competition_name: competition.name,
-          competition_slug: competition.slug,
-          division_count: items.length,
-        })
-        window.location.href = result.checkoutUrl
-        return
-      }
-
-      throw new Error("Failed to create checkout session")
-    } catch (err) {
-      const errorMessage =
-        err instanceof Error ? err.message : "Registration failed"
-
-      trackEvent("competition_registration_failed", {
-        competition_id: competition.id,
-        competition_name: competition.name,
-        competition_slug: competition.slug,
-        error_type: "unknown",
-      })
-
-      toast.error(errorMessage)
-      setIsSubmitting(false)
-    }
-  }
-
-  const formatDate = (date: string | Date | number | null): string => {
-    if (!date) return "TBA"
-
-    if (typeof date === "string") {
-      const match = date.match(/^(\d{4})-(\d{2})-(\d{2})$/)
-      if (match) {
-        const [, yearStr, monthStr, dayStr] = match
-        const year = Number(yearStr)
-        const monthNum = Number(monthStr)
-        const day = Number(dayStr)
-        const d = new Date(Date.UTC(year, monthNum - 1, day))
-        const weekdays = [
-          "Sunday",
-          "Monday",
-          "Tuesday",
-          "Wednesday",
-          "Thursday",
-          "Friday",
-          "Saturday",
-        ]
-        const months = [
-          "January",
-          "February",
-          "March",
-          "April",
-          "May",
-          "June",
-          "July",
-          "August",
-          "September",
-          "October",
-          "November",
-          "December",
-        ]
-        return `${weekdays[d.getUTCDay()]}, ${months[monthNum - 1]} ${day}, ${year}`
-      }
-      return "TBA"
-    }
-
-    const d = typeof date === "number" ? new Date(date) : date
-    return d.toLocaleDateString("en-US", {
-      weekday: "long",
-      month: "long",
-      day: "numeric",
-      year: "numeric",
-    })
-  }
-
-  const getRegistrationMessage = () => {
-    if (!registrationOpensAt || !registrationClosesAt) {
-      return "Registration dates have not been set yet."
-    }
-
-    const now = new Date()
-    const todayStr = getLocalDateKey(now)
-
-    if (todayStr < registrationOpensAt) {
-      return `Registration opens ${formatDate(registrationOpensAt)} and closes ${formatDate(registrationClosesAt)}`
-    }
-    if (todayStr > registrationClosesAt) {
-      return `Registration was open from ${formatDate(registrationOpensAt)} to ${formatDate(registrationClosesAt)}`
-    }
-    return null
-  }
-
-  const registrationMessage = getRegistrationMessage()
-  const hasSelectedDivisions = selectedDivisionIds.length > 0
-
-  // Determine if submit should be disabled
-  const competitionFull = competitionCapacity?.isFull ?? false
+  const competitionFull = props.competitionCapacity?.isFull ?? false
   const submitDisabled =
-    isSubmitting ||
-    !registrationOpen ||
+    r.isSubmitting ||
+    !props.registrationOpen ||
     competitionFull ||
-    !hasSelectedDivisions ||
-    !affiliateName.trim() ||
-    (waivers.length > 0 && !allRequiredWaiversAgreed)
+    !r.hasSelectedDivisions ||
+    !r.affiliateName.trim() ||
+    (props.waivers.length > 0 && !r.allRequiredWaiversAgreed)
 
-  // Get selected team divisions (for rendering team fields)
-  const selectedTeamDivisions = selectedDivisionIds
-    .map((id) => getDivision(id))
-    .filter((d): d is ScalingLevel => d !== undefined && d.teamSize > 1)
+  const fieldsDisabled = r.isSubmitting || !props.registrationOpen
 
   return (
     <div className="space-y-6">
-      <div className="space-y-2">
-        <h1 className="text-3xl font-bold">Register for Competition</h1>
-        <p className="text-muted-foreground">{competition.name}</p>
-      </div>
+      <PageHeader competition={props.competition} />
+      <CapacityBanners capacity={props.competitionCapacity} />
+      <RemovedAlert
+        removedDivisionIds={props.removedDivisionIds ?? []}
+        getDivision={r.getDivision}
+      />
+      <RegisteredAlert
+        registeredDivisionIds={props.registeredDivisionIds ?? []}
+      />
+      {!props.registrationOpen ? (
+        <ClosedRegistrationBanner
+          registrationOpensAt={props.registrationOpensAt}
+          registrationClosesAt={props.registrationClosesAt}
+        />
+      ) : null}
+      <CompetitionDetailsCard
+        competition={props.competition}
+        registrationOpensAt={props.registrationOpensAt}
+        registrationClosesAt={props.registrationClosesAt}
+      />
 
-      {competitionCapacity?.isFull && (
-        <Card className="border-amber-500/50 bg-amber-500/10">
-          <CardContent className="pt-6">
-            <div className="flex items-start gap-2 text-amber-600">
-              <AlertTriangle className="h-5 w-5 mt-0.5 shrink-0" />
-              <div>
-                <p className="font-semibold">Competition is full</p>
-                <p className="text-sm text-muted-foreground">
-                  This competition has reached its maximum number of registrations.
-                </p>
-              </div>
-            </div>
-          </CardContent>
-        </Card>
-      )}
-
-      {competitionCapacity && !competitionCapacity.isFull && competitionCapacity.spotsAvailable !== null && competitionCapacity.spotsAvailable <= 5 && (
-        <Card className="border-amber-500/50 bg-amber-500/10">
-          <CardContent className="pt-6">
-            <div className="flex items-center gap-2 text-amber-600">
-              <Users className="h-4 w-4 shrink-0" />
-              <p className="text-sm font-semibold">
-                Only {competitionCapacity.spotsAvailable} spot{competitionCapacity.spotsAvailable === 1 ? '' : 's'} left across all divisions!
-              </p>
-            </div>
-          </CardContent>
-        </Card>
-      )}
-
-      {removedDivisionIds.length > 0 && (
-        <Card className="border-destructive/20 bg-destructive/5">
-          <CardContent className="pt-6">
-            <div className="flex items-start gap-2 text-destructive">
-              <AlertTriangle className="h-4 w-4 mt-0.5 shrink-0" />
-              <div className="text-sm space-y-1">
-                <p className="font-medium">
-                  Your registration
-                  {removedDivisionIds.length === 1 ? " was" : "s were"} removed
-                  from this competition:
-                </p>
-                <ul className="list-disc pl-4">
-                  {removedDivisionIds.map((id) => {
-                    const div = getDivision(id)
-                    return <li key={id}>{div?.label ?? "Unknown division"}</li>
-                  })}
-                </ul>
-                <p className="text-muted-foreground">
-                  If you believe this was a mistake, please contact the event
-                  organizer.
-                </p>
-              </div>
-            </div>
-          </CardContent>
-        </Card>
-      )}
-
-      {registeredDivisionIds.length > 0 && (
-        <Card className="border-green-500/20 bg-green-500/5">
-          <CardContent className="pt-6">
-            <div className="flex items-center gap-2 text-green-600 dark:text-green-400">
-              <CheckCircle2 className="h-4 w-4" />
-              <span className="text-sm font-medium">
-                You're already registered for{" "}
-                {registeredDivisionIds.length === 1
-                  ? "1 division"
-                  : `${registeredDivisionIds.length} divisions`}
-                . Select additional divisions below.
-              </span>
-            </div>
-          </CardContent>
-        </Card>
-      )}
-
-      {!registrationOpen && registrationMessage && (
-        <Card className="border-yellow-500/50 bg-yellow-500/10">
-          <CardContent className="pt-6">
-            <p className="text-sm font-medium">{registrationMessage}</p>
-          </CardContent>
-        </Card>
-      )}
-
-      <Card>
-        <CardHeader>
-          <CardTitle>Competition Details</CardTitle>
-        </CardHeader>
-        <CardContent className="space-y-2">
-          <div>
-            <p className="text-muted-foreground text-sm">
-              {isSameDateString(competition.startDate, competition.endDate)
-                ? "Competition Date"
-                : "Competition Dates"}
-            </p>
-            <p className="font-medium">
-              {isSameDateString(competition.startDate, competition.endDate)
-                ? formatDate(competition.startDate)
-                : `${formatDate(competition.startDate)} - ${formatDate(competition.endDate)}`}
-            </p>
-          </div>
-          <div>
-            <p className="text-muted-foreground text-sm">Registration Window</p>
-            <p className="font-medium">
-              {formatDate(registrationOpensAt)} -{" "}
-              {formatDate(registrationClosesAt)}
-            </p>
-          </div>
-          <div>
-            <p className="text-muted-foreground text-sm">Hosted By</p>
-            <p className="font-medium">
-              {competition.organizingTeam?.name || "TBA"}
-            </p>
-          </div>
-        </CardContent>
-      </Card>
-
-      <form onSubmit={onSubmit} className="space-y-6">
-        {/* Division Selection - Searchable multi-select dropdown */}
-        <Card>
-          <CardHeader>
-            <CardTitle>
-              Select Division{selectedDivisionIds.length > 1 ? "s" : ""}
-            </CardTitle>
-            <CardDescription>
-              Choose one or more divisions to register for
-            </CardDescription>
-          </CardHeader>
-          <CardContent>
-            <DivisionMultiSelect
-              scalingLevels={scalingGroup.scalingLevels}
-              publicDivisions={publicDivisions}
-              selectedIds={selectedDivisionIds}
-              registeredDivisionIds={registeredDivisionIdSet}
-              removedDivisionIds={removedDivisionIdSet}
-              onToggle={handleDivisionToggle}
-              disabled={isSubmitting || !registrationOpen || competitionFull}
-            />
-            {/* Selected divisions shown as badges */}
-            {hasSelectedDivisions && (
-              <div className="flex flex-wrap gap-1.5 mt-3">
-                {selectedDivisionIds.map((id) => {
-                  const level = getDivision(id)
-                  if (!level) return null
-                  return (
-                    <Badge
-                      key={id}
-                      variant="secondary"
-                      className="pl-2 pr-1 py-1 gap-1"
-                    >
-                      {level.label}
-                      <button
-                        type="button"
-                        onClick={() => handleDivisionToggle(id, false)}
-                        className="ml-0.5 rounded-full hover:bg-muted-foreground/20 p-0.5"
-                        disabled={isSubmitting || !registrationOpen}
-                      >
-                        <X className="w-3 h-3" />
-                      </button>
-                    </Badge>
-                  )
-                })}
-              </div>
-            )}
-            {!hasSelectedDivisions && (
-              <p className="text-sm text-muted-foreground mt-3">
-                Select at least one division to continue
-              </p>
-            )}
-          </CardContent>
-        </Card>
-
-        {/* Affiliate Selection Card */}
-        <Card>
-          <CardHeader>
-            <CardTitle>Your Affiliate</CardTitle>
-            <CardDescription>
-              Select your gym or affiliate. Choose "Independent" if you don't
-              train at a gym.
-            </CardDescription>
-          </CardHeader>
-          <CardContent>
-            <div className="space-y-2">
-              <Label>Affiliate *</Label>
-              <AffiliateCombobox
-                value={affiliateName}
-                onChange={setAffiliateName}
-                placeholder="Search or select affiliate..."
-                disabled={isSubmitting || !registrationOpen}
-              />
-              <p className="text-xs text-muted-foreground">
-                Your gym or affiliate name will be displayed on leaderboards
-              </p>
-            </div>
-          </CardContent>
-        </Card>
-
-        {/* Registration Questions Card */}
-        {questions.length > 0 && (
-          <Card>
-            <CardHeader>
-              <CardTitle>Registration Questions</CardTitle>
-            </CardHeader>
-            <CardContent className="space-y-6">
-              {questions.map((question) => {
-                const answer = answers.find((a) => a.questionId === question.id)
-                return (
-                  <div key={question.id} className="space-y-2">
-                    <Label>
-                      {question.label}
-                      {question.required && (
-                        <span className="text-destructive"> *</span>
-                      )}
-                    </Label>
-                    {question.type === "select" ? (
-                      <Select
-                        onValueChange={(val) => updateAnswer(question.id, val)}
-                        value={answer?.answer || undefined}
-                        disabled={isSubmitting || !registrationOpen}
-                      >
-                        <SelectTrigger>
-                          <SelectValue placeholder="Select an option" />
-                        </SelectTrigger>
-                        <SelectContent>
-                          {question.options?.map((opt) => (
-                            <SelectItem key={opt} value={opt}>
-                              {opt}
-                            </SelectItem>
-                          ))}
-                        </SelectContent>
-                      </Select>
-                    ) : question.type === "number" ? (
-                      <Input
-                        type="number"
-                        value={answer?.answer ?? ""}
-                        onChange={(e) =>
-                          updateAnswer(question.id, e.target.value)
-                        }
-                        disabled={isSubmitting || !registrationOpen}
-                      />
-                    ) : (
-                      <Input
-                        value={answer?.answer ?? ""}
-                        onChange={(e) =>
-                          updateAnswer(question.id, e.target.value)
-                        }
-                        disabled={isSubmitting || !registrationOpen}
-                      />
-                    )}
-                    {question.helpText && (
-                      <p className="text-xs text-muted-foreground">
-                        {question.helpText}
-                      </p>
-                    )}
-                  </div>
-                )
-              })}
-            </CardContent>
-          </Card>
-        )}
-
-        {/* Registration Fee Card */}
-        {hasSelectedDivisions && (
-          <Card>
-            <CardHeader>
-              <CardTitle>
-                Registration Fee{selectedDivisionIds.length > 1 ? "s" : ""}
-              </CardTitle>
-            </CardHeader>
-            <CardContent className="space-y-3">
-              {selectedDivisionIds.map((divisionId) => {
-                const division = getDivision(divisionId)
-                const isMulti = selectedDivisionIds.length > 1
-                const hideDivTotal = isMulti || !!activeCoupon
-                return (
-                  <div key={divisionId}>
-                    {isMulti && (
-                      <p className="text-sm font-medium mb-1">
-                        {division?.label ?? "Division"}
-                      </p>
-                    )}
-                    <FeeBreakdown
-                      competitionId={competition.id}
-                      divisionId={divisionId}
-                      hideTotal={hideDivTotal}
-                      onFeesLoaded={handleFeesLoaded}
-                    />
-                  </div>
-                )
-              })}
-              {divisionFees.size > 0 && (() => {
-                const subtotal = Array.from(divisionFees.values()).reduce(
-                  (sum, c) => sum + c,
-                  0,
-                )
-                if (!activeCoupon) {
-                  // No coupon — show simple total for multi-division
-                  if (selectedDivisionIds.length <= 1) return null
-                  return (
-                    <div className="flex justify-between font-medium pt-2 border-t">
-                      <span>Total</span>
-                      <span className="text-lg">
-                        ${(subtotal / 100).toFixed(2)}
-                      </span>
-                    </div>
-                  )
-                }
-                // With coupon — always show subtotal, discount, and adjusted total
-                const discount = Math.min(activeCoupon.amountOffCents, subtotal)
-                const total = subtotal - discount
-                return (
-                  <>
-                    {selectedDivisionIds.length > 1 && (
-                      <div className="flex justify-between text-sm pt-2 border-t">
-                        <span>Subtotal</span>
-                        <span>${(subtotal / 100).toFixed(2)}</span>
-                      </div>
-                    )}
-                    <div className={cn(
-                      "flex justify-between text-sm text-emerald-700 dark:text-emerald-400",
-                      selectedDivisionIds.length <= 1 && "pt-2 border-t",
-                    )}>
-                      <span className="flex items-center gap-1.5">
-                        <Tag className="h-3.5 w-3.5" />
-                        Coupon ({activeCoupon.code})
-                      </span>
-                      <span>-${(discount / 100).toFixed(2)}</span>
-                    </div>
-                    <div className="flex justify-between font-medium pt-2 border-t">
-                      <span>Total</span>
-                      <span className="text-lg">
-                        ${(total / 100).toFixed(2)}
-                      </span>
-                    </div>
-                  </>
-                )
-              })()}
-            </CardContent>
-          </Card>
-        )}
-
-        {/* Team Registration Fields - one section per team division */}
-        {selectedTeamDivisions.map((division) => {
-          const teamEntry = teamEntries.get(division.id)
-          if (!teamEntry) return null
-
-          return (
-            <Card key={division.id}>
-              <CardHeader>
-                <CardTitle>
-                  Team Details
-                  {selectedTeamDivisions.length > 1 && (
-                    <span className="text-muted-foreground font-normal">
-                      {" "}
-                      - {division.label}
-                    </span>
-                  )}
-                </CardTitle>
-                <CardDescription>
-                  Enter your team name and invite your teammates for{" "}
-                  {division.label} ({division.teamSize} athletes)
-                </CardDescription>
-              </CardHeader>
-              <CardContent className="space-y-6">
-                <div className="space-y-2">
-                  <Label>Team Name *</Label>
-                  <Input
-                    placeholder="Enter your team name"
-                    value={teamEntry.teamName}
-                    onChange={(e) =>
-                      updateTeamEntry(division.id, "teamName", e.target.value)
-                    }
-                    disabled={isSubmitting || !registrationOpen}
-                  />
-                  <p className="text-xs text-muted-foreground">
-                    This will be displayed on leaderboards
-                  </p>
-                </div>
-
-                {/* Teammates */}
-                <div className="space-y-4">
-                  <div className="flex items-center justify-between">
-                    <h4 className="text-sm font-medium">Teammates</h4>
-                    <Badge variant="outline">
-                      {teamEntry.teammates.length + 1} of {division.teamSize}{" "}
-                      added
-                    </Badge>
-                  </div>
-
-                  {/* Current user as Teammate 1 (read-only) */}
-                  <Card className="p-4 bg-muted/50">
-                    <div className="space-y-4">
-                      <div className="flex items-center gap-2">
-                        <User className="w-4 h-4 text-muted-foreground" />
-                        <span className="font-medium">Teammate 1 (You)</span>
-                      </div>
-                      <div className="grid grid-cols-2 gap-4">
-                        <div>
-                          <Label className="text-sm font-medium">
-                            First Name
-                          </Label>
-                          <Input
-                            value={userFirstName || ""}
-                            disabled
-                            className="mt-1.5"
-                          />
-                        </div>
-                        <div>
-                          <Label className="text-sm font-medium">
-                            Last Name
-                          </Label>
-                          <Input
-                            value={userLastName || ""}
-                            disabled
-                            className="mt-1.5"
-                          />
-                        </div>
-                      </div>
-                      <div>
-                        <Label className="text-sm font-medium">Email</Label>
-                        <Input
-                          value={userEmail || ""}
-                          disabled
-                          className="mt-1.5"
-                        />
-                      </div>
-                      <p className="text-xs text-muted-foreground">
-                        Update your profile to change this information.
-                      </p>
-                    </div>
-                  </Card>
-
-                  {teamEntry.teammates.map((teammate, index) => (
-                    <Card
-                      key={`${division.id}-teammate-${index}`}
-                      className="p-4"
-                    >
-                      <div className="space-y-4">
-                        <div className="flex items-center gap-2">
-                          <Users className="w-4 h-4 text-muted-foreground" />
-                          <span className="font-medium">
-                            Teammate {index + 2}
-                          </span>
-                        </div>
-
-                        <div className="space-y-2">
-                          <Label>Email *</Label>
-                          <Input
-                            type="email"
-                            placeholder="teammate@email.com"
-                            value={teammate.email}
-                            onChange={(e) =>
-                              updateTeammate(
-                                division.id,
-                                index,
-                                "email",
-                                e.target.value,
-                              )
-                            }
-                            disabled={isSubmitting || !registrationOpen}
-                          />
-                        </div>
-
-                        <div className="grid grid-cols-2 gap-4">
-                          <div className="space-y-2">
-                            <Label>First Name</Label>
-                            <Input
-                              placeholder="First name"
-                              value={teammate.firstName}
-                              onChange={(e) =>
-                                updateTeammate(
-                                  division.id,
-                                  index,
-                                  "firstName",
-                                  e.target.value,
-                                )
-                              }
-                              disabled={isSubmitting || !registrationOpen}
-                            />
-                          </div>
-                          <div className="space-y-2">
-                            <Label>Last Name</Label>
-                            <Input
-                              placeholder="Last name"
-                              value={teammate.lastName}
-                              onChange={(e) =>
-                                updateTeammate(
-                                  division.id,
-                                  index,
-                                  "lastName",
-                                  e.target.value,
-                                )
-                              }
-                              disabled={isSubmitting || !registrationOpen}
-                            />
-                          </div>
-                        </div>
-
-                        <div className="space-y-2">
-                          <Label>Affiliate (Optional)</Label>
-                          <AffiliateCombobox
-                            value={teammate.affiliateName}
-                            onChange={(val) =>
-                              updateTeammate(
-                                division.id,
-                                index,
-                                "affiliateName",
-                                val,
-                              )
-                            }
-                            placeholder="Search or enter affiliate..."
-                            disabled={isSubmitting || !registrationOpen}
-                          />
-                        </div>
-                      </div>
-                    </Card>
-                  ))}
-
-                  <p className="text-sm text-muted-foreground">
-                    Teammates will receive an email invitation to join your
-                    team. They must accept the invite to complete their
-                    registration.
-                  </p>
-                </div>
-              </CardContent>
-            </Card>
-          )
-        })}
-
-        {/* Waivers Section */}
-        {waivers.length > 0 && (
-          <Card>
-            <CardHeader>
-              <CardTitle>Waivers & Agreements</CardTitle>
-              <CardDescription>
-                Please review and agree to the following waivers to complete
-                your registration
-              </CardDescription>
-            </CardHeader>
-            <CardContent className="space-y-6">
-              {waivers.map((waiver) => (
-                <div key={waiver.id} className="space-y-4">
-                  <div className="flex items-center gap-2">
-                    <h4 className="font-medium">{waiver.title}</h4>
-                    {waiver.required && (
-                      <Badge variant="destructive" className="text-xs">
-                        Required
-                      </Badge>
-                    )}
-                  </div>
-
-                  <div className="border rounded-lg p-4 max-h-64 overflow-y-auto bg-muted/10">
-                    <WaiverViewer
-                      content={waiver.content}
-                      className="prose prose-sm max-w-none dark:prose-invert"
-                    />
-                  </div>
-
-                  <div className="flex items-start gap-3 p-4 bg-muted/20 rounded-lg">
-                    <Checkbox
-                      id={`waiver-${waiver.id}`}
-                      checked={agreedWaivers.has(waiver.id)}
-                      onCheckedChange={(checked) =>
-                        handleWaiverCheckChange(waiver.id, checked === true)
-                      }
-                      disabled={isSubmitting || !registrationOpen}
-                    />
-                    <Label
-                      htmlFor={`waiver-${waiver.id}`}
-                      className="text-sm font-medium leading-none cursor-pointer"
-                    >
-                      I have read and agree to this waiver
-                      {waiver.required && (
-                        <span className="text-destructive ml-1">*</span>
-                      )}
-                    </Label>
-                  </div>
-                </div>
-              ))}
-            </CardContent>
-          </Card>
-        )}
-
+      <form onSubmit={r.onSubmit} className="space-y-6">
+        <DivisionPickerSection
+          scalingLevels={props.scalingGroup.scalingLevels}
+          publicDivisions={props.publicDivisions}
+          selectedIds={r.selectedDivisionIds}
+          registeredDivisionIds={r.registeredDivisionIdSet}
+          removedDivisionIds={r.removedDivisionIdSet}
+          getDivision={r.getDivision}
+          onToggle={r.handleDivisionToggle}
+          disabled={fieldsDisabled || competitionFull}
+        />
+        <AffiliateSection
+          value={r.affiliateName}
+          onChange={r.setAffiliateName}
+          disabled={fieldsDisabled}
+        />
+        <RegistrationQuestionsSection
+          questions={props.questions}
+          answers={r.answers}
+          onAnswerChange={r.updateAnswer}
+          disabled={fieldsDisabled}
+        />
+        <FeeSummarySection
+          competitionId={props.competition.id}
+          selectedDivisionIds={r.selectedDivisionIds}
+          getDivision={r.getDivision}
+          divisionFees={r.divisionFees}
+          onFeesLoaded={r.handleFeesLoaded}
+          activeCoupon={r.activeCoupon}
+        />
+        <TeamDetailsSection
+          selectedTeamDivisions={r.selectedTeamDivisions}
+          teamEntries={r.teamEntries}
+          updateTeamEntry={r.updateTeamEntry}
+          updateTeammate={r.updateTeammate}
+          userFirstName={props.userFirstName}
+          userLastName={props.userLastName}
+          userEmail={props.userEmail}
+          disabled={fieldsDisabled}
+        />
+        <WaiversSection
+          waivers={props.waivers}
+          agreedWaivers={r.agreedWaivers}
+          onWaiverToggle={r.handleWaiverCheckChange}
+          disabled={fieldsDisabled}
+        />
         <div className="flex gap-4">
           <Button type="submit" disabled={submitDisabled} className="flex-1">
-            {isSubmitting ? (
+            {r.isSubmitting ? (
               <>
                 <Loader2 className="w-4 h-4 mr-2 animate-spin" />
                 Processing...
               </>
-            ) : !registrationOpen ? (
+            ) : !props.registrationOpen ? (
               "Registration Closed"
             ) : competitionFull ? (
               "Competition Full"
-            ) : !hasSelectedDivisions ? (
+            ) : !r.hasSelectedDivisions ? (
               "Select a Division"
-            ) : waivers.length > 0 && !allRequiredWaiversAgreed ? (
+            ) : props.waivers.length > 0 && !r.allRequiredWaiversAgreed ? (
               "Agree to Waivers to Continue"
-            ) : selectedDivisionIds.length > 1 ? (
-              `Register for ${selectedDivisionIds.length} Divisions`
-            ) : selectedTeamDivisions.length > 0 ? (
+            ) : r.selectedDivisionIds.length > 1 ? (
+              `Register for ${r.selectedDivisionIds.length} Divisions`
+            ) : r.selectedTeamDivisions.length > 0 ? (
               "Register Team"
             ) : (
               "Complete Registration"
@@ -1340,8 +174,10 @@ export function RegistrationForm({
           <Button
             type="button"
             variant="outline"
-            onClick={() => navigate({ to: `/compete/${competition.slug}` })}
-            disabled={isSubmitting}
+            onClick={() =>
+              navigate({ to: `/compete/${props.competition.slug}` })
+            }
+            disabled={r.isSubmitting}
           >
             Cancel
           </Button>
@@ -1350,3 +186,115 @@ export function RegistrationForm({
     </div>
   )
 }
+
+// ─── Invite variant ─────────────────────────────────────────────────────────
+// The invitation IS the authorization: bypasses the public registration window,
+// pins the division, and never shows a "Registration Closed" state.
+
+export function InviteRegistrationForm(props: InviteProps) {
+  const r = useRegistrationForm(props)
+  const navigate = useNavigate()
+
+  // If the URL pointed at a division that's no longer eligible (full,
+  // already registered, removed), fall back to the public flow so the
+  // athlete can pick a different division rather than seeing nothing.
+  if (!r.invitedDivision) {
+    const {
+      initialDivisionId: _ignored,
+      inviteToken: __ignored,
+      ...rest
+    } = props
+    return <PublicRegistrationForm {...rest} registrationOpen={true} />
+  }
+
+  const submitDisabled =
+    r.isSubmitting ||
+    !r.affiliateName.trim() ||
+    (props.waivers.length > 0 && !r.allRequiredWaiversAgreed)
+  const fieldsDisabled = r.isSubmitting
+
+  return (
+    <div className="space-y-6">
+      <PageHeader competition={props.competition} />
+      <CapacityBanners capacity={props.competitionCapacity} />
+      <CompetitionDetailsCard
+        competition={props.competition}
+        registrationOpensAt={props.registrationOpensAt}
+        registrationClosesAt={props.registrationClosesAt}
+      />
+      <InviteDivisionHero
+        championshipName={props.competition.name}
+        divisionLabel={r.invitedDivision.label}
+        teamSize={r.invitedDivision.teamSize ?? 1}
+        imageUrl={props.competition.profileImageUrl}
+      />
+
+      <form onSubmit={r.onSubmit} className="space-y-6">
+        <AffiliateSection
+          value={r.affiliateName}
+          onChange={r.setAffiliateName}
+          disabled={fieldsDisabled}
+        />
+        <RegistrationQuestionsSection
+          questions={props.questions}
+          answers={r.answers}
+          onAnswerChange={r.updateAnswer}
+          disabled={fieldsDisabled}
+        />
+        <FeeSummarySection
+          competitionId={props.competition.id}
+          selectedDivisionIds={r.selectedDivisionIds}
+          getDivision={r.getDivision}
+          divisionFees={r.divisionFees}
+          onFeesLoaded={r.handleFeesLoaded}
+          activeCoupon={r.activeCoupon}
+        />
+        <TeamDetailsSection
+          selectedTeamDivisions={r.selectedTeamDivisions}
+          teamEntries={r.teamEntries}
+          updateTeamEntry={r.updateTeamEntry}
+          updateTeammate={r.updateTeammate}
+          userFirstName={props.userFirstName}
+          userLastName={props.userLastName}
+          userEmail={props.userEmail}
+          disabled={fieldsDisabled}
+        />
+        <WaiversSection
+          waivers={props.waivers}
+          agreedWaivers={r.agreedWaivers}
+          onWaiverToggle={r.handleWaiverCheckChange}
+          disabled={fieldsDisabled}
+        />
+        <div className="flex gap-4">
+          <Button type="submit" disabled={submitDisabled} className="flex-1">
+            {r.isSubmitting ? (
+              <>
+                <Loader2 className="w-4 h-4 mr-2 animate-spin" />
+                Processing...
+              </>
+            ) : props.waivers.length > 0 && !r.allRequiredWaiversAgreed ? (
+              "Agree to Waivers to Continue"
+            ) : r.selectedTeamDivisions.length > 0 ? (
+              "Confirm Team Spot"
+            ) : (
+              "Confirm Spot"
+            )}
+          </Button>
+          <Button
+            type="button"
+            variant="outline"
+            onClick={() =>
+              navigate({ to: `/compete/${props.competition.slug}` })
+            }
+            disabled={r.isSubmitting}
+          >
+            Cancel
+          </Button>
+        </div>
+      </form>
+    </div>
+  )
+}
+
+// Re-export the hook return type for any caller that wants it.
+export type { UseRegistrationFormReturn }
