+++ apps/wodsmith-start/src/routes/compete/$slug/register.tsx
@@ -10,7 +10,10 @@ import { createFileRoute, notFound, redirect } from "@tanstack/react-router"
 import { createServerFn } from "@tanstack/react-start"
 import { and, eq, inArray, isNotNull } from "drizzle-orm"
 import { z } from "zod"
-import { RegistrationForm } from "@/components/registration/registration-form"
+import {
+  InviteRegistrationForm,
+  PublicRegistrationForm,
+} from "@/components/registration/registration-form"
 import {
   competitionRegistrationAnswersTable,
   competitionRegistrationsTable,
@@ -312,15 +315,17 @@ export const Route = createFileRoute("/compete/$slug/register")({
 
     // 6. Get scaling group and levels for divisions (via server function)
     // Also get public divisions for capacity info
-    const [{ scalingGroup }, { divisions: publicDivisions, competitionCapacity }] =
-      await Promise.all([
-        getScalingGroupWithLevelsFn({
-          data: { scalingGroupId: settings.divisions.scalingGroupId },
-        }),
-        getPublicCompetitionDivisionsFn({
-          data: { competitionId: competition.id },
-        }),
-      ])
+    const [
+      { scalingGroup },
+      { divisions: publicDivisions, competitionCapacity },
+    ] = await Promise.all([
+      getScalingGroupWithLevelsFn({
+        data: { scalingGroupId: settings.divisions.scalingGroupId },
+      }),
+      getPublicCompetitionDivisionsFn({
+        data: { competitionId: competition.id },
+      }),
+    ])
 
     if (
       !scalingGroup ||
@@ -395,7 +400,11 @@ function RegisterPage() {
     signedWaiverIds,
   } = Route.useLoaderData()
 
-  const { canceled } = Route.useSearch()
+  const {
+    canceled,
+    divisionId: initialDivisionId,
+    invite: inviteToken,
+  } = Route.useSearch()
 
   // Show error if divisions are not configured
   if (!divisionsConfigured || !scalingGroup) {
@@ -415,29 +424,47 @@ function RegisterPage() {
     )
   }
 
+  // Variant dispatch: presence of `?divisionId=...` (the only producer is the
+  // claim CTA) routes the athlete into the invite-aware variant. The server
+  // is the actual authority — it looks up the active pending invite for
+  // (session email, competition, division) and bypasses the public window if
+  // one exists, regardless of whether the URL also carried `?invite=<token>`.
+  // The token is decoration; the database invite row is the source of truth.
+  const sharedProps = {
+    competition,
+    scalingGroup,
+    publicDivisions,
+    competitionCapacity,
+    userId,
+    registrationOpensAt,
+    registrationClosesAt,
+    paymentCanceled: canceled === "true",
+    defaultAffiliateName,
+    waivers,
+    questions,
+    userFirstName,
+    userLastName,
+    userEmail,
+    registeredDivisionIds,
+    removedDivisionIds,
+    previousAnswers,
+    signedWaiverIds,
+  }
+
   return (
     <div className="mx-auto max-w-2xl">
-      <RegistrationForm
-        competition={competition}
-        scalingGroup={scalingGroup}
-        publicDivisions={publicDivisions}
-        competitionCapacity={competitionCapacity}
-        userId={userId}
-        registrationOpen={registrationOpen}
-        registrationOpensAt={registrationOpensAt}
-        registrationClosesAt={registrationClosesAt}
-        paymentCanceled={canceled === "true"}
-        defaultAffiliateName={defaultAffiliateName}
-        waivers={waivers}
-        questions={questions}
-        userFirstName={userFirstName}
-        userLastName={userLastName}
-        userEmail={userEmail}
-        registeredDivisionIds={registeredDivisionIds}
-        removedDivisionIds={removedDivisionIds}
-        previousAnswers={previousAnswers}
-        signedWaiverIds={signedWaiverIds}
-      />
+      {initialDivisionId ? (
+        <InviteRegistrationForm
+          {...sharedProps}
+          initialDivisionId={initialDivisionId}
+          inviteToken={inviteToken}
+        />
+      ) : (
+        <PublicRegistrationForm
+          {...sharedProps}
+          registrationOpen={registrationOpen}
+        />
+      )}
     </div>
   )
 }
