+++ lat.md/competition-invites.md
@@ -100,6 +100,10 @@ Phase 2 sub-arc B wires three pages: `$token.tsx` (claim landing), `$token/decli
 
 `$token.tsx` is the entry point clicked from the email. The loader calls [[apps/wodsmith-start/src/server-fns/competition-invite-fns.ts#getInviteByTokenFn]] — a public (session-free) server fn that resolves the token, confirms the invite's championship slug matches the URL (anti-typo guard), and returns either `{ kind: "not_claimable", reason }` or `{ kind: "claimable", invite, championshipName, divisionLabel, accountExistsForInviteEmail }`. The loader then runs `identityMatch` against `context.session`. The four-way branch is: happy-path → render pre-attached registration CTA; `wrong_account` → render "this invite is for a different account" page; `needs_sign_in` / `needs_sign_up` → `redirect` into `/sign-in?redirect=…&email=<invite.email>&invite=<token>` (or sign-up equivalent) so post-auth re-entry lands back on the claim page with a session that matches.
 
+The happy-path CTA links to `/compete/$slug/register?divisionId=<invite.championshipDivisionId>`. The register route's search schema accepts `divisionId` (and a Phase 2D `invite=<token>`). Presence of `divisionId` routes the athlete to [[apps/wodsmith-start/src/components/registration/registration-form.tsx#InviteRegistrationForm]] instead of [[apps/wodsmith-start/src/components/registration/registration-form.tsx#PublicRegistrationForm]] — explicit variants rather than mode booleans. Both variants share state through [[apps/wodsmith-start/src/components/registration/use-registration-form.ts#useRegistrationForm]] and compose the same section components ([[apps/wodsmith-start/src/components/registration/registration-sections.tsx#AffiliateSection]], etc.).
+
+The invite variant renders [[apps/wodsmith-start/src/components/registration/invite-division-hero.tsx#InviteDivisionHero]] in place of the picker, pins the invited division as the only selection, and bypasses the public registration window — the invitation itself is the authorization to register. If the URL points at a division that's no longer eligible (full, already-registered, removed) the variant transparently falls back to the public flow so the athlete can still pick something else rather than seeing nothing.
+
 `$token/decline.tsx` uses the same loader branches, then POSTs through [[apps/wodsmith-start/src/server-fns/competition-invite-fns.ts#declineInviteFn]] on confirmation. The server fn re-runs resolve + identity-match (server is the authority — a forged POST can't bypass email-lock) before calling [[apps/wodsmith-start/src/server/competition-invites/decline.ts#declineInvite]], which transitions `pending → declined` and nulls `activeMarker` + `claimTokenHash` so the link dies immediately. Declines are idempotent at the DB level: a zero-affected-rows outcome re-reads to tell "already terminal" apart from "row missing".
 
 `invite-pending.tsx` is the landing that wrong-account visitors bounce to after signing out and back in — it tells them to re-open the email rather than guessing a URL. The claim route's wrong-account page also links here via `/sign-in?redirect=/compete/$slug/invite-pending&email=<invite.email>`.
